update: GrandMA2 Lua Plugin Readme.md docs
Update the structure in the 'Readme.md' doc for the 'Check Cue Existence' plugin.
Update the structure in the 'Readme.md' doc for the 'Cue Sequence Generator' plugin.
Update the structure in the 'Readme.md' doc for the 'Execute Cues at Random' plugin.
Update the structure in the 'Readme.md' doc for the 'Get Last Object' plugin.